Divide and conquer client side funcs
