+++ client/client.go
@@ -1,7 +1,9 @@
 package client
 
 import (
+	"bufio"
 	"context"
+	"fmt"
 
 	"github.com/h0n9/petit-chat/msg"
 	"github.com/h0n9/petit-chat/p2p"
@@ -84,4 +86,59 @@ func (c *Client) LeaveMsgBox(topicStr string) error {
 
 func (c *Client) GetMsgBox(topicStr string) (*msg.Box, bool) {
 	return c.msgCenter.GetBox(topicStr)
+}
+
+func (cli *Client) StartChat(topic string, reader *bufio.Reader) error {
+	box, exist := cli.GetMsgBox(topic)
+	if !exist {
+		fmt.Printf("Type nickname: ")
+		nickname, err := util.GetInput(reader, false, false)
+		if err != nil {
+			return err
+		}
+		fmt.Printf("Type public('true', 't' or 'false', 'f'): ")
+		pubStr, err := util.GetInput(reader, false, true)
+		if err != nil {
+			return err
+		}
+		pub, err := util.ToBool(pubStr)
+		if err != nil {
+			return err
+		}
+		b, err := cli.CreateMsgBox(topic, nickname, pub)
+		if err != nil {
+			return err
+		}
+		box = b
+	}
+
+	chat, err := NewChat(box, reader)
+	if err != nil {
+		return err
+	}
+	defer chat.Close()
+
+	// open subscription
+	go box.Subscribe(msg.DefaultMsgHandler)
+
+	// get and print out received msgs
+	msgs := box.GetUnreadMsgs()
+	for _, msg := range msgs {
+		err := readMsg(box, msg)
+		if err != nil {
+			fmt.Printf("%s\n> ", err)
+		}
+	}
+
+	// get and print out new msgs
+	chat.wg.Add(1)
+	go chat.output()
+
+	// get user input
+	chat.wg.Add(1)
+	go chat.input()
+
+	chat.wg.Wait()
+
+	return nil
 }
